Step 9 of refactoring the fight mechanic - a lot of changes done here to League and Contender classes as well as minor changes to the Blue and Red robo classes which are now renamed to YogiBear and BooBoo contenders. Full details listed below:
(League.java)
MAX_ROUNDS variable added to limit the fights to this number, set to 200 (tested with Yogi Bear vs. BooBoo fight in which BooBoo wins by K.O. after 199 rounds)
roundCount added to while loop for fight, breaks out of fight if roundCount hits MAX_ROUNDS
whole fight mechanic updated to the new rules of attack or defend each round, this creates 4 possible outcomes of a round 1. both attack, 2. A attacks and B defends, 3. A defends and B attacks and 4. both defend
roundCount used in outcome determination to add new possibilities of "jury decision" when MAX_ROUNDS is reached the remaining health of the players is used to determine the winner
NOTE: A known limitation of the "jury decision" happens if both players defend and reach MAX rounds Contender B is declared the winner, the solution for this would be to introduce another variable to track damage delt as a 2nd metric for announcing a winner, at this point it remains random as Contender A vs B is selected randomly.
added additional initialisation of Contender Joe and Bob in case no contenders are around - use this also to test the "jury decision" draw scenario

(Contender.java)
MAX_HEALTH variable added to be used in health heal cap - so Contender can heal up past 100
energy variable added and set to 40
energyDrain variable added, set to 0 by default but updated in the new energyCheck() method
hit(int damage) method updated to run the new energyCheck() method and added output to console
energyCheck() method added to either damage or heal the Contender based on energy usage, works on H = (E<40) ? H + (40 - E) : H - (E - 40) where 40 is the MAX_ENERGY
AI() method added to have automatic behaviour of attack then defend then restart and repeat

(BlueRobo.java -> renamed to -> YogiBear.java)
Yogi Bear modified to have a defensePower of 25 and a weaker attack of 15
@Override added to the AI() method to make Yogi Bear always attack each round

(RedRobo.java -> renamed to -> BooBoo.java)
BooBoo modified to have a weaker attack of 1 to help test the healing mechanic